First version of ROS2 remote
Some tuning might still be needed.

What is the unit of received speed from ROS2? Need to scale to hoverboard speed?

What is the unit of firmware realSpeed? RPM?
Not so important as velocity is just published for information, i.e. not used ros2_control.
ros2_control only use wheelX_cnt as feedback and calculate odom.